refactor(backlog_core): extract full-text search engine into FastMCP-free search module
